+++ CHANGES.rst
@@ -19,3 +19,4 @@ v0.1.2
 
 v0.1.3[unreleased]
 -----------
+* Add address model